All source code
This assessment system involves two users: a teacher and a student.
Here, teachers can consult existing classes, create, edit and delete questions and tests and finally consult student statistics.

Students, on the other hand, have the possibility to solve the tests associated with the class they belong to and consult their own stats.